stacks improvement

+++ Microbench/src/main/scala/benchmarks/Step.scala
@@ -5,15 +5,15 @@ import org.openjdk.jmh.annotations._
 @State(Scope.Thread)
 class Step {
   @Param(Array("2"))
-  var step: Long = _
+  var step: Int = _
 
-  var count: Long = 0
+  var count: Int = 0
 
-  def run(): Long = {
+  def run(): Int = {
     count += 1
     count
   }
-  def get(): Long = count
-  def at(): Long = step
-  def test(v: Long): Boolean = v % step == 0
+  def get(): Int = count
+  def at(): Int = step
+  def test(v: Int): Boolean = v % step == 0
 }
